requirements.txt
python-dotenv


langchain
openai
cohere
ai21
pygpt4all
huggingface_hub